Change loadbalancer type for HCP deployments
The Kubevirt(HCP) clusters do not have their loadbalancer installedand rely
on the host loadblancer( like netallb) when a loadbalancer is created.So when
the Submariner exposes the submariner-gateway loadbalancer service with
ServiceExternalTrafficPolicy locally does not work, as the traffic from
the remotecluster first lands on the host cluster. Hence the traffic policy
is updated to clusterin these deployments.

Signed-off-by: Aswin Suryanarayanan <asuryana@redhat.com>