Switch to Go 1.24 tool handling
Instead of specifying the tool dependencies in a Go file, this relies
on Go 1.24's tool handling. Tools are still managed in a separate
module to avoid intermingling dependencies.

Instead of direct "go tool" invocations (which don't work well with a
separate module, unless an undocumented parameter is used) or local
cached binaries (which makes it hard to handle rebuilds when tool
versions change on systems with pre-existing versions of the tools),
this uses a gotool macro to retrieve the cached binary used by "go
tool". "go tool -n" is invoked twice because the first invocation
returns the path to the initial binary, which is a temporary binary
removed as soon as "go tool" completes; the second invocation returns
the path to the cached binary.

lichen is only used once, so it is handled inline; controller-gen and
kustomize are referenced multiple times, so they are handled using
variables to store the paths, as before. operator-sdk is more
complicated to build, so the existing mechanism is used; the only
change there is that the tool dependency is represented using "go
tool" semantics in tools/go.mod.

Signed-off-by: Stephen Kitt <skitt@redhat.com>