Skip JSON error checking when unnecessary
It is possible to determine statically that the argument to
json.Marshal or json.MarshalIndent is safe to encode, and can't
produce an error. In such cases there is no point in checking the
returned error. As a side benefit, removing the "if err" blocks
reduces measured cyclomatic complexity and increases code coverage.

errchkjson can detect this, but even when enabled, its checks default
to disabled. This enables the checks and configures errcheck to ignore
json.Marshal and json.MarshalIndent (this is safe because they are
checked by errchkjson).

Signed-off-by: Stephen Kitt <skitt@redhat.com>